Move registry cmd from snowpark to top level
