+++ src/snowcli/cli/registry/commands.py
@@ -2,7 +2,7 @@ import typer
 
 from snowcli.cli.common.decorators import global_options_with_connection
 from snowcli.cli.common.flags import DEFAULT_CONTEXT_SETTINGS, ConnectionOption
-from snowcli.cli.snowpark.registry.manager import get_token
+from snowcli.cli.registry.manager import get_token
 from snowcli.output.decorators import with_output
 from snowcli.output.types import ObjectResult
