+++ src/snowcli/cli/registry/manager.py
