+++ src/snowcli/cli/snowpark/__init__.py
@@ -9,7 +9,6 @@ from snowcli.cli.snowpark.compute_pool.commands import (
 )
 from snowcli.cli.snowpark.services.commands import app as services_app
 from snowcli.cli.snowpark.jobs.commands import app as jobs_app
-from snowcli.cli.snowpark.registry.commands import app as registry_app
 
 app = typer.Typer(
     name="snowpark",
@@ -23,4 +22,3 @@ app.add_typer(procedure_app)  # type: ignore
 app.add_typer(compute_pools_app)  # type: ignore
 app.add_typer(services_app)  # type: ignore
 app.add_typer(jobs_app)  # type: ignore
-app.add_typer(registry_app)  # type: ignore
